requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
pandas==2.2.3
openai==1.46.0
httpx==0.27.2
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-auth-httplib2==0.2.0
gspread==6.1.4
jinja2==3.1.4
python-multipart==0.0.9
XlsxWriter==3.2.0
numpy>=1.24.0
